Updated clusters, improved lsh

+++ Point_Table_Implementation.cpp
@@ -1,6 +1,7 @@
 
 
 #include "Point_Table.hpp"
+#include "Hypercube.hpp"
 
 #include "utilities.hpp"
 #include <string>
@@ -149,12 +150,12 @@ int Point_Array::Compute_g(int position,int k,int M, const long long int m, doub
 
 }
 
-
+/*
 string Point_Array::Compute_f(int position, int k, int M, const long long int m, 
 					double w, double** s_params, Hypercube* hcube){
 
 	string concat = "";
-	long long h;
+	int h;
 	char bin;
 
     //Run input and find LSH_Manhattan k times
@@ -174,6 +175,7 @@ string Point_Array::Compute_f(int position, int k, int M, const long long int m,
 	cout << "bitstring is" << concat;
 	return concat;
 }
+*/
 
 
 /* Returning the image in position */
